Redirecting after user creates new profile
