+++ communio/lib/view/Widgets/create_profile_form.dart
@@ -1,18 +1,14 @@
 import 'dart:convert';
 import 'dart:io';
-import 'package:communio/redux/action_creators.dart';
+import 'package:communio/redux/actions.dart';
 
 import 'package:communio/model/app_state.dart';
-import 'package:communio/model/known_person.dart';
 import 'package:communio/view/Widgets/image_upload.dart';
 import 'package:communio/view/Widgets/insert_email_field.dart';
 import 'package:communio/view/Widgets/insert_name_field.dart';
 import 'package:communio/view/Widgets/insert_password_field.dart';
-import 'package:communio/view/Widgets/profile_interests.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
-
-import 'package:logger/logger.dart';
 import 'package:toast/toast.dart';
 import 'package:http/http.dart' as http;
 
@@ -78,7 +74,7 @@ class CreateProfileFormState extends State<CreateProfileForm> {
     }
 
     if (_formKey.currentState.validate() && _agreedWithTerms()) {
-      Toast.show('Processing Data', context);
+      Toast.show('Processing Data', context, duration: 3);
       final name = nameController.text.trim();
       final email = emailController.text.trim();
       final password = passwordController.text.trim();
@@ -95,13 +91,18 @@ class CreateProfileFormState extends State<CreateProfileForm> {
         "password": password,
       });
 
-      var response = this.addPerson(context, req);
-      if (response != 200) {
-        Toast.show('Invalid values!', context, duration: 5);
-        return;
-      }
+      this.addPerson(context, req).then((response) {
+        if (response['status'] != 200) {
+          Toast.show(response['msg'], context, duration: 10);
+          //return;
+        }
 
-      StoreProvider.of<AppState>(context).dispatch(updateUser(email));
+        getUser(context, email).then((value) {
+            StoreProvider.of<AppState>(context).dispatch(UpdateUser(value));
+            Navigator.of(context).pushNamed('/Homepage');
+        });
+
+      });
     }
   }
 
@@ -132,7 +133,21 @@ class CreateProfileFormState extends State<CreateProfileForm> {
                 ))));
   }
 
-  addPerson(BuildContext context, reg) async {
+  Future<String> getUser(BuildContext context, String email) async {
+    final store = StoreProvider.of<AppState>(context);
+    final response = await http.get('${store.state.content['profile']}');
+    var map = json.decode(utf8.decode(response.bodyBytes));
+    var id;
+    for (var item in map) {
+      if (item['email'] == email) {
+        id = item['_id'];
+        break;
+      }
+    }
+    return id;
+  }
+
+  Future<Map<String,Object>> addPerson(BuildContext context, reg) async {
     final store = StoreProvider.of<AppState>(context);
     Map<String, String> headers = {
       'Content-type': 'application/json',
@@ -143,7 +158,7 @@ class CreateProfileFormState extends State<CreateProfileForm> {
         body: reg, headers: headers);
     print(utf8.decode(response.bodyBytes));
     print(response.statusCode);
-    return response.statusCode;
+    return {"status": response.statusCode, "msg": utf8.decode(response.bodyBytes)};
   }
 
   bool _agreedWithTerms() {
